src/pod42-server/requirements.txt
langchain
openai 
pinecone-client
transformers
tiktoken
sanic
